+++ pkg/odo/cli/component/create.go
@@ -10,7 +10,6 @@ import (
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 	"github.com/zalando/go-keyring"
-	"k8s.io/klog"
 
 	devfilev1 "github.com/devfile/api/pkg/apis/workspaces/v1alpha2"
 	"github.com/devfile/library/pkg/devfile"
@@ -28,11 +27,9 @@ import (
 	"github.com/openshift/odo/pkg/log"
 	"github.com/openshift/odo/pkg/machineoutput"
 	appCmd "github.com/openshift/odo/pkg/odo/cli/application"
-	catalogutil "github.com/openshift/odo/pkg/odo/cli/catalog/util"
 	"github.com/openshift/odo/pkg/odo/cli/component/ui"
 	projectCmd "github.com/openshift/odo/pkg/odo/cli/project"
 	registryUtil "github.com/openshift/odo/pkg/odo/cli/registry/util"
-	commonui "github.com/openshift/odo/pkg/odo/cli/ui"
 	"github.com/openshift/odo/pkg/odo/genericclioptions"
 	odoutil "github.com/openshift/odo/pkg/odo/util"
 	"github.com/openshift/odo/pkg/odo/util/completion"
@@ -40,7 +37,6 @@ import (
 	"github.com/openshift/odo/pkg/preference"
 	"github.com/openshift/odo/pkg/util"
 
-	corev1 "k8s.io/api/core/v1"
 	ktemplates "k8s.io/kubectl/pkg/util/templates"
 )
 
@@ -370,8 +366,15 @@ func (co *CreateOptions) Complete(name string, cmd *cobra.Command, args []string
 	}
 	co.DevfilePath = DevfilePath
 
-	if util.CheckPathExists(ConfigFilePath) || util.CheckPathExists(EnvFilePath) {
+	if util.CheckPathExists(ConfigFilePath) {
 		return errors.New("this directory already contains a component")
+	}
+
+	if util.CheckPathExists(EnvFilePath) && !util.CheckPathExists(co.DevfilePath) {
+		log.Warningf("Found a dangling env file without a devfile, overriding it")
+		if err := util.DeletePath(EnvFilePath); err != nil {
+			return err
+		}
 	}
 
 	if util.CheckPathExists(co.DevfilePath) && co.devfileMetadata.devfilePath.value != "" && !util.PathEqual(co.DevfilePath, co.devfileMetadata.devfilePath.value) {
@@ -670,160 +673,21 @@ func (co *CreateOptions) Complete(name string, cmd *cobra.Command, args []string
 
 	// Below code is for INTERACTIVE mode
 	if co.interactive {
-		client := co.Client
-
 		if len(catalogList.Items) == 0 {
-			catalogList, err = catalog.ListComponents(client)
+			catalogList, err = catalog.ListComponents(co.Client)
 			if err != nil {
 				return err
 			}
 		}
 
-		componentTypeCandidates := catalogutil.FilterHiddenComponents(catalogList.Items)
-		selectedComponentType := ui.SelectComponentType(componentTypeCandidates)
-		selectedImageTag := ui.SelectImageTag(componentTypeCandidates, selectedComponentType)
-		componentType := selectedComponentType + ":" + selectedImageTag
-		co.componentSettings.Type = &componentType
-
-		// Ask for the type of source if not provided
-		selectedSourceType := ui.SelectSourceType([]config.SrcType{config.LOCAL, config.GIT, config.BINARY})
-		co.componentSettings.SourceType = &selectedSourceType
-		selectedSourcePath := LocalDirectoryDefaultLocation
-
-		// Get the current directory
-		currentDirectory, err := os.Getwd()
+		err := co.SetComponentSettingsInteractively(catalogList)
 		if err != nil {
 			return err
 		}
-
-		if selectedSourceType == config.BINARY {
-
-			// We ask for the source of the component context
-			co.componentContext = ui.EnterInputTypePath("context", currentDirectory, ".")
-			klog.V(4).Infof("Context: %s", co.componentContext)
-
-			// If it's a binary, we have to ask where the actual binary in relation
-			// to the context
-			selectedSourcePath = ui.EnterInputTypePath("binary", ".")
-
-			// Get the correct source location
-			sourceLocation, err := getSourceLocation(selectedSourcePath, co.componentContext)
-			if err != nil {
-				return errors.Wrapf(err, "unable to get source location")
-			}
-			co.componentSettings.SourceLocation = &sourceLocation
-
-		} else if selectedSourceType == config.GIT {
-
-			// For git, we ask for the Git URL and set that as the source location
-			cmpSrcLOC, selectedGitRef := ui.EnterGitInfo()
-			co.componentSettings.SourceLocation = &cmpSrcLOC
-			co.componentSettings.Ref = &selectedGitRef
-
-		} else if selectedSourceType == config.LOCAL {
-
-			// We ask for the source of the component, in this case the "path"!
-			co.componentContext = ui.EnterInputTypePath("path", currentDirectory, ".")
-
-			// Get the correct source location
-			if co.componentContext == "" {
-				co.componentContext = LocalDirectoryDefaultLocation
-			}
-			co.componentSettings.SourceLocation = &co.componentContext
-
-		}
-
-		defaultComponentName, err := createDefaultComponentName(co.Context, selectedComponentType, selectedSourceType, selectedSourcePath)
-		if err != nil {
-			return err
-		}
-		componentName := ui.EnterComponentName(defaultComponentName, co.Context)
-
-		appName := ui.EnterOpenshiftName(co.Context.Application, "Which application do you want the component to be associated with", co.Context)
-		co.componentSettings.Application = &appName
-
-		projectName := ui.EnterOpenshiftName(co.Context.Project, "Which project go you want the component to be created in", co.Context)
-		co.componentSettings.Project = &projectName
-
-		co.componentSettings.Name = &componentName
-
-		var ports []string
-
-		if commonui.Proceed("Do you wish to set advanced options") {
-			// if the user doesn't opt for advanced options, ports field would remain unpopulated
-			// we then set it at the end of this function
-			ports = ui.EnterPorts()
-
-			co.componentEnvVars = ui.EnterEnvVars()
-
-			if commonui.Proceed("Do you wish to set resource limits") {
-				memMax := ui.EnterMemory("maximum", "512Mi")
-				memMin := ui.EnterMemory("minimum", memMax)
-				cpuMax := ui.EnterCPU("maximum", "1")
-				cpuMin := ui.EnterCPU("minimum", cpuMax)
-
-				memoryQuantity, err := util.FetchResourceQuantity(corev1.ResourceMemory, memMin, memMax, "")
-				if err != nil {
-					return err
-				}
-				if memoryQuantity != nil {
-					co.componentSettings.MinMemory = &memMin
-					co.componentSettings.MaxMemory = &memMax
-				}
-				cpuQuantity, err := util.FetchResourceQuantity(corev1.ResourceCPU, cpuMin, cpuMax, "")
-				if err != nil {
-					return err
-				}
-				if cpuQuantity != nil {
-					co.componentSettings.MinCPU = &cpuMin
-					co.componentSettings.MaxCPU = &cpuMax
-				}
-			}
-		}
-
-		// if user didn't opt for advanced options, "ports" value remains empty which panics the "odo push"
-		// so we set the ports here
-		if len(ports) == 0 {
-			ports, err = co.Client.GetPortsFromBuilderImage(*co.componentSettings.Type)
-			if err != nil {
-				return err
-			}
-		}
-
-		co.componentSettings.Ports = &ports
-		// Above code is for INTERACTIVE mode
-
 	} else {
-		// Else if NOT using interactive / UI
-		err = co.setComponentSourceAttributes()
-		if err != nil {
-			return err
-		}
-		err = co.setComponentName(args)
-		if err != nil {
-			return err
-		}
-
-		var portList []string
-		if len(co.componentPorts) > 0 {
-			portList = co.componentPorts
-		} else {
-			portList, err = co.Client.GetPortsFromBuilderImage(*co.componentSettings.Type)
-			if err != nil {
-				return err
-			}
-		}
-
-		co.componentSettings.Ports = &(portList)
+		co.SetComponentSettings(args)
 	}
 
-	co.componentSettings.Project = &(co.Context.Project)
-	envs, err := config.NewEnvVarListFromSlice(co.componentEnvVars)
-	if err != nil {
-		return
-	}
-	co.componentSettings.Envs = envs
-	co.ignores = []string{}
 	if co.now {
 		co.ResolveSrcAndConfigFlags()
 		err = co.ResolveProject(co.Context.Project)
@@ -1037,9 +901,8 @@ func (co *CreateOptions) downloadStarterProject(devObj parser.DevfileObj, projec
 
 	} else if starterProject.Zip != nil {
 		url := starterProject.Zip.Location
-		logUrl := starterProject.Zip.Location
 		sparseDir := starterProject.SubDir
-		downloadSpinner := log.Spinnerf("Downloading starter project %s from %s", starterProject.Name, logUrl)
+		downloadSpinner := log.Spinnerf("Downloading starter project %s from %s", starterProject.Name, url)
 		err := co.checkoutProject(sparseDir, url, path)
 		if err != nil {
 			downloadSpinner.End(false)
@@ -1246,27 +1109,6 @@ func (co *CreateOptions) checkoutProject(subDir, zipURL, path string) error {
 	return nil
 }
 
-// getStarterProjectInteractiveMode gets starter project value by asking user in interactive mode.
-func getStarterProjectInteractiveMode(projects []devfilev1.StarterProject) *devfilev1.StarterProject {
-	projectName := ui.SelectStarterProject(projects)
-
-	// if user do not wish to download starter project or there are no projects in devfile, project name would be empty
-	if projectName == "" {
-		return nil
-	}
-
-	var project devfilev1.StarterProject
-
-	for _, value := range projects {
-		if value.Name == projectName {
-			project = value
-			break
-		}
-	}
-
-	return &project
-}
-
 // getStarterProjectFromFlag gets starter project value from flag --starter.
 func getStarterProjectFromFlag(projects []devfilev1.StarterProject, projectPassed string) (project *devfilev1.StarterProject, err error) {
 
@@ -1299,6 +1141,37 @@ func getStarterProjectFromFlag(projects []devfilev1.StarterProject, projectPasse
 
 	return project, err
 
+}
+
+func (co *CreateOptions) SetComponentSettings(args []string) error {
+	err := co.setComponentSourceAttributes()
+	if err != nil {
+		return err
+	}
+	err = co.setComponentName(args)
+	if err != nil {
+		return err
+	}
+
+	var portList []string
+	if len(co.componentPorts) > 0 {
+		portList = co.componentPorts
+	} else {
+		portList, err = co.Client.GetPortsFromBuilderImage(*co.componentSettings.Type)
+		if err != nil {
+			return err
+		}
+	}
+
+	co.componentSettings.Ports = &(portList)
+	co.componentSettings.Project = &(co.Context.Project)
+	envs, err := config.NewEnvVarListFromSlice(co.componentEnvVars)
+	if err != nil {
+		return err
+	}
+	co.componentSettings.Envs = envs
+	co.ignores = []string{}
+	return nil
 }
 
 // NewCmdCreate implements the create odo command
