+++ tests/integration/devfile/cmd_devfile_create_test.go
@@ -214,10 +214,17 @@ var _ = Describe("odo devfile create command tests", func() {
 	})
 
 	Context("When executing odo create with component with no devBuild command", func() {
-		It("should successfully create the devfile component", func() {
+		It("should successfully create the devfile component and remove a dangling env file", func() {
 			// Quarkus devfile has no devBuild command
 			output := helper.CmdShouldPass("odo", "create", "java-quarkus")
 			helper.MatchAllInOutput(output, []string{"Please use `odo push` command to create the component with source deployed"})
+			helper.DeleteFile("devfile.yaml")
+			out, outerr := helper.CmdShouldPassIncludeErrStream("odo", "create", "nodejs")
+			helper.MatchAllInOutput(out, []string{
+				"Please use `odo push` command to create the component with source deployed"})
+			helper.MatchAllInOutput(outerr, []string{
+				"Found a dangling env file without a devfile, overriding it",
+			})
 		})
 	})
